[UPDATE] add function for FPS

+++ math/convolution.hpp
@@ -56,7 +56,7 @@ custom_complex<T> pol(T r, T theta){
   return custom_complex<T>(r * cosl(theta), r * sinl(theta));
 }
 
-using cd = custom_complex<double>;
+using cd = custom_complex<long double>;
 // f.size() should be the power of 2.
 void rec_fft(vector<cd> &f, bool inv=false){
   int n = f.size();
@@ -82,7 +82,7 @@ void fft(vector<cd> &f, bool inv=false){
   int n = f.size(), mask = n - 1;
   vector<cd> tmp(n);
   for(int i = n >> 1; i >= 1; i >>= 1){
-    cd zeta = pol<double>(1., 2. * m_pi * i * (inv ? -1.: 1.) / n);
+    cd zeta = pol<long double>(1., 2. * m_pi * i * (inv ? -1.: 1.) / n);
     cd w = cd(1., 0);
     for(int j = 0; j < n; j += i){
       for(int k = 0; k < i; k++){
